requirements.txt
aiofiles==23.2.1
grpcio==1.59.3
grpcio-tools==1.59.3
numpy==1.26.2
pandas==2.1.3
pgvector==0.2.3
protobuf==4.25.1
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg-pool==3.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==69.0.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
